Add star layer
This commit adds support for the `Star rating` layer. This mean that we had to
pull in a menu package to render a menu for picking the layer. It still needs a
lot of styling!

We also made the toolbar buttons bigger and changed the `today` button to an up
angle icon, which still looks bad…