Enable double-fingered drag on map element
Actually this is a bit of a hack, and not at all as nice as the google maps one
which shows a helper overlay about double-fingered drag. But it does allow the
UI to be scrolled, so we can get access to the possibly hidden UI elements while
we sort out 100vh properly.