Fix iOS portrait mode
Finally! This seems reasonable. It's still not perfect since you're stuck with
the useless Safari toolbars, but if you install it as an app, then you get rid
of those.

Landscape mode is still weirdly broken: you get a scrollbar and black space
below the map and button bar. No idea why! But we're having other issues with
the menu, so we need to fix that as well before calling landscape mode 'done'.