Use image proxy to work around http(s)
RASP images are server only over `http`. Trying to load an http image in an
https react app seems to result in the protocol being rewritten to https - I
suspect this is the browser attempting https first and then going to fall back
to http. But in RASP's case the https request hangs for a long time, which I
suspect breaks the browser behaviour :(

Anyhow, there seems to be a free image proxy in NL (images.weserv.nl), so we'll
try using that. It seems to add a little latency to the app :(

Hopefully RASP will add https support at some point. They are aware of it being
broken, and realise that browsers will be unhappy with http traffic in the
future, so are motivated to fix it.